functionMails/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions==1.7.2
pkg_resources==0.0.0
psycopg2-binary==2.9.1
python-http-client==3.3.3
sendgrid==6.3.0
